requirements.txt

python-dotenv==1.1.1
pandas==2.3.3